@@ -1,29 +1,7 @@
-[tool.isort]
-profile = "black"
-line_length = 120
-skip = ["venv", ".tox", "docs"]
-known_third_party = ["pydantic", "migrations"]
-known_first_party = ["test", "docs"]
-
-[tool.black]
-line-length = 120
-target-version = ["py310"]
-exclude = '''
-(
-  /(
-    geant_service_orchestrator\.egg-info      # exclude a few common directories in the
-    | \.git                                   # root of the project
-    | \.*_cache
-    | \.tox
-    | venv
-    | docs
-  )/
-)
-'''
-
 [tool.mypy]
 exclude = [
     "venv",
+    "test/*",
     "docs"
 ]
 ignore_missing_imports = true
@@ -44,64 +22,81 @@ show_error_codes = true
 show_column_numbers = true
 # Suppress "note: By default the bodies of untyped functions are not checked"
 disable_error_code = "annotation-unchecked"
+# Forbid the use of a generic "type: ignore" without specifying the exact error that is ignored
+enable_error_code = "ignore-without-code"
 
 [tool.ruff]
-exclude = [
-    ".git",
-    ".*_cache",
-    ".tox",
-    "*.egg-info",
-    "__pycache__",
+extend-exclude = [
     "htmlcov",
-    "venv",
-    "docs"
+    "docs",
 ]
 ignore = [
-    "C417",
-    "D100",
-    "D101",
-    "D102",
-    "D103",
-    "D104",
-    "D105",
-    "D106",
-    "D107",
-    "D202",
     "D203",
     "D213",
-    "E501",
-    "N806",
-    "B905",
     "N805",
-    "B904",
-    "N803",
-    "N801",
-    "N815",
-    "N802",
-    "S101",
-    "S104"
+    "PLR0913",
+    "PLR0904",
+    "PLW1514"
 ]
 line-length = 120
 select = [
+    "A",
+    "ARG",
     "B",
+    "BLE",
     "C",
+    "COM",
+    "C4",
+    "C90",
     "D",
+    "DTZ",
     "E",
+    "EM",
+    "ERA",
     "F",
+    "FA",
+    "FBT",
+    "FLY",
+    "FURB",
+    "G",
     "I",
+    "ICN",
+    "INP",
+    "ISC",
+    "LOG",
     "N",
+    "PERF",
+    "PGH",
+    "PIE",
+    "PL",
+    "PT",
+    "PTH",
+    "PYI",
+    "Q",
     "RET",
+    "R",
+    "RET",
+    "RSE",
+    "RUF",
     "S",
+    "SIM",
+    "SLF",
     "T",
+    "T20",
+    "TID",
+    "TRY",
+    "UP",
     "W",
+    "YTT"
 ]
-target-version = "py310"
+target-version = "py311"
 
 [tool.ruff.flake8-tidy-imports]
 ban-relative-imports = "all"
 
 [tool.ruff.per-file-ignores]
-"test/*" = ["B033", "N816", "N802"]
+"test/*" = ["D", "S101", "PLR2004"]
+"setup.py" = ["D100"]
 
 [tool.ruff.isort]
 known-third-party = ["pydantic", "migrations"]
